# Conflicts:
#	CMakeLists.txt
#	src/Mutect2/AssemblyResultSet.h
#	src/Mutect2Engine.cpp
#	src/Mutect2Engine.h
#	src/haplotype/Haplotype.cpp
#	src/haplotype/Haplotype.h
#	src/haplotypecaller/AssemblyBasedCallerUtils.cpp
#	src/main.cpp
#	src/variantcontext/VariantContext.h

+++ src/graph/KBestHaplotype.cpp
@@ -14,7 +14,8 @@ KBestHaplotype::KBestHaplotype(std::shared_ptr<SeqVertex> initialVertex,
 
 KBestHaplotype::KBestHaplotype(const std::shared_ptr<KBestHaplotype> &p, const std::shared_ptr<BaseEdge> &edge,
                                int totalOutgoingMultiplicity) : Path<SeqVertex, BaseEdge>(*p, edge) {
-	score = p->getScore() + std::log10(edge->getMultiplicity()) - std::log10(totalOutgoingMultiplicity);
+	score = p->getScore() + (double) (std::log10((long double) edge->getMultiplicity()) -
+	                                  std::log10((long double) totalOutgoingMultiplicity));
 	isReference &= edge->getIsRef();
 }
